+++ apps/docs/src/content/04-components/navigation/tab-navigation/examples/default.tsx
@@ -1,0 +1,19 @@
+import {
+  TabNavigation,
+  Link,
+  AlertIcon,
+} from "@mittwald/flow-react-components";
+
+<TabNavigation aria-label="Projekt-Navigation">
+  <Link href="#">Apps</Link>
+  <Link href="#" aria-current="page">
+    Container
+  </Link>
+  <Link href="#">Domains</Link>
+  <Link href="#">E-Mails</Link>
+  <Link href="#">
+    Datenbanken
+    <AlertIcon status="warning" />
+  </Link>
+  <Link href="#">Backups</Link>
+</TabNavigation>;