+++ apps/docs/src/content/04-components/structure/tabs/examples/default-selected.tsx
@@ -1,0 +1,16 @@
+import {
+  Tab,
+  Tabs,
+  TabTitle,
+} from "@mittwald/flow-react-components";
+
+<Tabs defaultSelectedKey="container">
+  <Tab>
+    <TabTitle>Mit App verbinden</TabTitle>
+    App Content
+  </Tab>
+  <Tab id="container">
+    <TabTitle>Mit Container verbinden</TabTitle>
+    Container Content
+  </Tab>
+</Tabs>;